requirements.txt
# this is the "requirements.txt" file...

#testing
pytest

# stocks:
pandas
python-dotenv
plotly

#web app
flask
requests

#production server
gunicorn
